requirements.txt
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.4
buildozer==0.32
certifi==2018.8.24
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
Cython==0.23
decorator==4.1.2
docutils==0.13.1
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
html5lib==1.0.1
idna==2.7
imutils==0.4.3
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==7.4.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kazoo==2.6.1
Kivy==1.9.2.dev0
Kivy-Garden==0.1.4
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.12.1
pandas==0.23.4
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.2.0
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
scandir==1.5
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
stevedore==1.20.0
subprocess32==3.2.7
TBB==0.1
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
urllib3==1.23
vboxapi==1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
